separatec containerwith state definitions
